Moved authentication logic into concern
