+++ config/routes.rb
@@ -2,7 +2,7 @@ Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
   root 'calendar#index'
 
-  get 'oauth2callback' => 'calendar#authenticate'
+  get 'oauth2callback' => 'authentication#authenticate'
 
   get 'google_index' => 'calendar#google_index'
